+++ next-docs/public/examples/iconButton/AsLink.tsx
@@ -1,12 +1,8 @@
 import { IconButton } from '@heathmont/moon-core-tw';
-import { GenericSettings } from '@heathmont/moon-icons-tw';
+import { OtherFrame } from '@heathmont/moon-icons-tw';
 
 const Example = () => (
-  <IconButton
-    as="a"
-    href="/core/button"
-    icon={<GenericSettings className="text-moon-24" />}
-  />
+  <IconButton as="a" href="/core/button" icon={<OtherFrame />} />
 );
 
 export default Example;