+++ next-docs/public/examples/iconButton/Disabled.tsx
@@ -1,8 +1,6 @@
 import { IconButton } from '@heathmont/moon-core-tw';
-import { GenericSettings } from '@heathmont/moon-icons-tw';
+import { OtherFrame } from '@heathmont/moon-icons-tw';
 
-const Example = () => (
-  <IconButton disabled icon={<GenericSettings className="text-moon-24" />} />
-);
+const Example = () => <IconButton disabled icon={<OtherFrame />} />;
 
 export default Example;