+++ next-docs/public/examples/iconButton/Sizes.tsx
@@ -1,13 +1,13 @@
 import { IconButton } from '@heathmont/moon-core-tw';
-import { GenericSettings } from '@heathmont/moon-icons-tw';
+import { OtherFrame } from '@heathmont/moon-icons-tw';
 
 const Example = () => (
   <>
-    <IconButton size="xs" icon={<GenericSettings className="text-moon-16" />} />
-    <IconButton size="sm" icon={<GenericSettings className="text-moon-24" />} />
-    <IconButton icon={<GenericSettings className="text-moon-24" />} />
-    <IconButton size="lg" icon={<GenericSettings className="text-moon-24" />} />
-    <IconButton size="xl" icon={<GenericSettings className="text-moon-24" />} />
+    <IconButton size="xs" icon={<OtherFrame />} />
+    <IconButton size="sm" icon={<OtherFrame />} />
+    <IconButton icon={<OtherFrame />} />
+    <IconButton size="lg" icon={<OtherFrame />} />
+    <IconButton size="xl" icon={<OtherFrame />} />
   </>
 );