+++ workspaces/core/src/button/private/utils/buttonSizes/getButtonSize.ts
@@ -16,25 +16,49 @@ const getButtonSize = ({
 }: ButtonSettingsProps): string => {
   if (size === 'xs') {
     return mergeClassnames(
-      getXsPadding({ icon, iconLeft, iconRight, iconOnly, fullWidth }),
+      getXsPadding({
+        icon,
+        iconLeft,
+        iconRight,
+        iconOnly,
+        fullWidth,
+      }),
       'h-6 py-1 gap-1 text-moon-12 rounded-moon-i-xs'
     );
   }
   if (size === 'sm') {
     return mergeClassnames(
-      getSmPadding({ icon, iconLeft, iconRight, iconOnly, fullWidth }),
+      getSmPadding({
+        icon,
+        iconLeft,
+        iconRight,
+        iconOnly,
+        fullWidth,
+      }),
       'h-8 py-1 gap-1 text-moon-14 rounded-moon-i-sm'
     );
   }
   if (size === 'lg') {
     return mergeClassnames(
-      getLgPadding({ icon, iconLeft, iconRight, iconOnly, fullWidth }),
+      getLgPadding({
+        icon,
+        iconLeft,
+        iconRight,
+        iconOnly,
+        fullWidth,
+      }),
       'h-12 py-3 gap-2 text-moon-16 rounded-moon-i-sm'
     );
   }
   if (size === 'xl') {
     return mergeClassnames(
-      getXlPadding({ icon, iconLeft, iconRight, iconOnly, fullWidth }),
+      getXlPadding({
+        icon,
+        iconLeft,
+        iconRight,
+        iconOnly,
+        fullWidth,
+      }),
       'h-14 py-4 gap-2 text-moon-16 rounded-moon-i-md'
     );
   }