+++ workspaces/core/src/button/private/utils/buttonStyles/getButtonCommonStyles.ts
@@ -1,9 +1,12 @@
 import mergeClassnames from '../../../../mergeClassnames/mergeClassnames';
 import type ButtonSettingsProps from '../../types/ButtonSettingsProps';
 
-const getButtonCommonStyles = ({ disabled }: ButtonSettingsProps): string =>
+const getButtonCommonStyles = (
+  disabled: ButtonSettingsProps['disabled']
+): string =>
   mergeClassnames(
-    'relative z-0 flex justify-center items-center font-medium no-underline overflow-hidden select-none outline-none transition duration-200 active:scale-90 focus:ring-4',
+    'relative z-0 flex justify-center items-center font-medium no-underline overflow-hidden',
+    'whitespace-nowrap select-none transition duration-200 active:scale-90',
     disabled && 'opacity-30 cursor-not-allowed'
   );