+++ workspaces/core/src/button/private/utils/buttonStyles/getButtonVariants.ts
@@ -1,16 +1,16 @@
 import type ButtonSettingsProps from '../../types/ButtonSettingsProps';
 
-const getButtonVariants = ({ variant }: ButtonSettingsProps): string => {
+const getButtonVariants = (variant: ButtonSettingsProps['variant']): string => {
   if (variant === 'secondary') {
-    return 'btn-secondary focus:ring-bulma/[.2]';
+    return 'btn-secondary';
   }
   if (variant === 'tertiary') {
-    return 'btn-tertiary focus:ring-piccolo/[.2]';
+    return 'btn-tertiary';
   }
   if (variant === 'ghost') {
-    return 'text-trunks bg-transparent hover:text-bulma focus:ring-bulma/[.2]';
+    return 'text-trunks bg-transparent hover:text-bulma';
   }
-  return 'btn-primary focus:ring-piccolo/[.2]';
+  return 'btn-primary';
 };
 
 export default getButtonVariants;