+++ workspaces/core/src/iconButton/styles/IconButtonComponent.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import getAnimation from '../../button/private/utils/buttonAnimations/getAnimation';
+import getIconSize from '../../button/private/utils/buttonSizes/getIconSize';
 import getButtonCommonStyles from '../../button/private/utils/buttonStyles/getButtonCommonStyles';
 import getButtonVariants from '../../button/private/utils/buttonStyles/getButtonVariants';
 import mergeClassnames from '../../mergeClassnames/mergeClassnames';
@@ -20,11 +21,12 @@ const IconButtonComponent = <C extends React.ElementType>({
   return (
     <Component
       className={mergeClassnames(
-        getButtonSize({ size }),
-        getButtonCommonStyles({ disabled }),
-        getButtonVariants({ variant }),
+        getButtonSize(size),
+        getButtonCommonStyles(disabled),
+        getButtonVariants(variant),
         animation === 'pulse' && getAnimation('pulse'),
         animation === 'error' && getAnimation('error'),
+        getIconSize(size),
         customClassName
       )}
       {...((!as || as === 'button') && { type: 'button' })}