+++ workspaces/core/src/popover/private/types/PopoverRootProps.ts
@@ -1,0 +1,7 @@
+import type Placement from './Placement';
+
+type PopoverRootProps = {
+  position?: Placement;
+};
+
+export default PopoverRootProps;