+++ workspaces/templates/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-templates
 
+## 9.18.0
+
+### Minor Changes
+
+- Deploy minor version
+
 ## 9.17.0
 
 ### Minor Changes